feat: Refactor navigation and home screen structure to support dynamic page indexing and add favorites view
